Allow a size of 0 in area's rectangle state
This makes it so that an area's rectangle state can have
rectangles with a width or height of 0, by allowing the top left corner to be identical to the bottom right instead of having to
be strictly smaller.

This makes it consistent with the other two states, as circles
are allowed to have a radius of 0 and polygons can have all
their points along a line or even in a single spot, making
areas with a size of 0 possible.

Fixes #10306.